+++ src/lib/types.rs
@@ -7,6 +7,7 @@
 #[path = "./types_test.rs"]
 mod types_test;
 
+use crate::legacy;
 use ci_info::types::CiInfo;
 use indexmap::IndexMap;
 use rust_info::types::RustInfo;
@@ -563,6 +564,8 @@ pub struct Task {
     /// if script exit code is not 0, the command/script of this task will not be invoked, dependencies however will be
     pub condition_script: Option<Vec<String>>,
     /// if true, any error while executing the task will be printed but will not break the build
+    pub ignore_errors: Option<bool>,
+    /// DEPRECATED, replaced with ignore_errors
     pub force: Option<bool>,
     /// The env vars to setup before running the task commands
     pub env: Option<IndexMap<String, EnvValue>>,
@@ -619,6 +622,7 @@ impl Task {
             watch: None,
             condition: None,
             condition_script: None,
+            ignore_errors: None,
             force: None,
             env: None,
             cwd: None,
@@ -704,6 +708,12 @@ impl Task {
             self.condition_script = task.condition_script.clone();
         } else if override_values {
             self.condition_script = None;
+        }
+
+        if task.ignore_errors.is_some() {
+            self.ignore_errors = task.ignore_errors.clone();
+        } else if override_values {
+            self.ignore_errors = None;
         }
 
         if task.force.is_some() {
@@ -833,9 +843,19 @@ impl Task {
         }
     }
 
-    /// Returns true if the task force attribute is defined and true
-    pub fn is_force(self: &Task) -> bool {
-        self.force.unwrap_or(false)
+    /// Returns true if the task ignore_errors attribute is defined and true
+    pub fn should_ignore_errors(self: &Task) -> bool {
+        match self.ignore_errors {
+            Some(value) => value,
+            None => match self.force {
+                Some(value) => {
+                    legacy::show_deprecated_attriute_warning("force", "ignore_errors");
+
+                    value
+                }
+                None => false,
+            },
+        }
     }
 
     /// Returns the override task definition based on the current platform.
@@ -875,6 +895,7 @@ impl Task {
                     watch: override_task.watch.clone(),
                     condition: override_task.condition.clone(),
                     condition_script: override_task.condition_script.clone(),
+                    ignore_errors: override_task.ignore_errors.clone(),
                     force: override_task.force.clone(),
                     env: override_task.env.clone(),
                     cwd: override_task.cwd.clone(),
@@ -968,6 +989,8 @@ pub struct PlatformOverrideTask {
     /// if script exit code is not 0, the command/script of this task will not be invoked, dependencies however will be
     pub condition_script: Option<Vec<String>>,
     /// if true, any error while executing the task will be printed but will not break the build
+    pub ignore_errors: Option<bool>,
+    /// DEPRECATED, replaced with ignore_errors
     pub force: Option<bool>,
     /// The env vars to setup before running the task commands
     pub env: Option<IndexMap<String, EnvValue>>,
@@ -1028,6 +1051,10 @@ impl PlatformOverrideTask {
 
             if self.condition_script.is_none() && task.condition_script.is_some() {
                 self.condition_script = task.condition_script.clone();
+            }
+
+            if self.ignore_errors.is_none() && task.ignore_errors.is_some() {
+                self.ignore_errors = task.ignore_errors.clone();
             }
 
             if self.force.is_none() && task.force.is_some() {