+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support additional watch options #188
 * Support glob paths in CARGO_MAKE_WORKSPACE_SKIP_MEMBERS #187
 * Add install-rust-src and install-rls to internal makefile #189
+* Renamed task 'force' attribute to 'ignore_errors' #120
 
 ### v0.16.7 (2019-01-25)